refactor: (#766) 알림 리팩터링
