+++ backend/src/main/java/com/votogether/domain/alarm/controller/AlarmQueryController.java
@@ -2,7 +2,8 @@ package com.votogether.domain.alarm.controller;
 
 import com.votogether.domain.alarm.dto.ReportActionAlarmResponse;
 import com.votogether.domain.alarm.dto.ReportActionResponse;
-import com.votogether.domain.alarm.service.AlarmService;
+import com.votogether.domain.alarm.dto.response.PostAlarmResponse;
+import com.votogether.domain.alarm.service.AlarmQueryService;
 import com.votogether.domain.member.entity.Member;
 import com.votogether.global.jwt.Auth;
 import jakarta.validation.constraints.PositiveOrZero;
@@ -10,6 +11,7 @@ import java.util.List;
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.ResponseEntity;
 import org.springframework.validation.annotation.Validated;
+import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestParam;
@@ -19,25 +21,33 @@ import org.springframework.web.bind.annotation.RestController;
 @RequiredArgsConstructor
 @RequestMapping("/alarms")
 @RestController
-public class AlarmCommandController implements AlarmCommandControllerDocs {
+public class AlarmQueryController implements AlarmQueryControllerDocs {
 
-    private final AlarmService alarmService;
+    private final AlarmQueryService alarmQueryService;
 
-    @RequestMapping("/report")
+    @GetMapping("/content")
+    public ResponseEntity<List<PostAlarmResponse>> getPostAlarm(
+            @RequestParam @PositiveOrZero(message = "페이지는 0이상 정수만 가능합니다.") final int page
+    ) {
+        final List<PostAlarmResponse> postAlarmResponses = alarmQueryService.getPostAlarm(page);
+        return ResponseEntity.ok(postAlarmResponses);
+    }
+
+    @GetMapping("/report")
     public ResponseEntity<List<ReportActionAlarmResponse>> getReportActionAlarms(
             @RequestParam @PositiveOrZero(message = "페이지는 0이상 정수만 가능합니다.") final int page,
             @Auth final Member member
     ) {
-        final List<ReportActionAlarmResponse> response = alarmService.getReportActionAlarms(member, page);
+        final List<ReportActionAlarmResponse> response = alarmQueryService.getReportActionAlarms(member, page);
         return ResponseEntity.ok(response);
     }
 
-    @RequestMapping("/report/{id}")
+    @GetMapping("/report/{id}")
     public ResponseEntity<ReportActionResponse> getReportActionAlarm(
             @PathVariable("id") final Long reportActionAlarmId,
             @Auth final Member member
     ) {
-        final ReportActionResponse response = alarmService.getReportActionAlarm(reportActionAlarmId, member);
+        final ReportActionResponse response = alarmQueryService.getReportActionAlarm(reportActionAlarmId, member);
         return ResponseEntity.ok(response);
     }
