+++ backend/src/main/java/com/votogether/domain/alarm/service/AlarmQueryService.java
@@ -1,14 +1,21 @@
 package com.votogether.domain.alarm.service;
 
+import com.votogether.domain.alarm.dto.ReportActionAlarmResponse;
+import com.votogether.domain.alarm.dto.ReportActionResponse;
 import com.votogether.domain.alarm.dto.response.PostAlarmResponse;
 import com.votogether.domain.alarm.entity.Alarm;
+import com.votogether.domain.alarm.entity.ReportActionAlarm;
 import com.votogether.domain.alarm.entity.vo.AlarmType;
+import com.votogether.domain.alarm.exception.ReportActionAlarmExceptionType;
 import com.votogether.domain.alarm.repository.AlarmRepository;
+import com.votogether.domain.alarm.repository.ReportActionAlarmRepository;
 import com.votogether.domain.member.entity.Member;
+import com.votogether.global.exception.NotFoundException;
 import java.util.List;
 import lombok.RequiredArgsConstructor;
 import org.springframework.data.domain.PageRequest;
 import org.springframework.data.domain.Slice;
+import org.springframework.data.domain.Sort;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
@@ -21,6 +28,7 @@ public class AlarmQueryService {
     private static final String NICKNAME_WHEN_POST_CLOSING = "";
 
     private final AlarmRepository alarmRepository;
+    private final ReportActionAlarmRepository reportActionAlarmRepository;
 
     public List<PostAlarmResponse> getPostAlarm(final int page) {
         final PageRequest pageRequest = PageRequest.of(page, BASIC_PAGE_SIZE);
@@ -44,6 +52,25 @@ public class AlarmQueryService {
         }
         final Member member = alarm.getMember();
         return member.getNickname();
+    }
+
+    public List<ReportActionAlarmResponse> getReportActionAlarms(final Member member, final int page) {
+        final PageRequest pageRequest = PageRequest.of(page, BASIC_PAGE_SIZE,
+                Sort.by(Sort.Direction.DESC, "createdAt"));
+        final List<ReportActionAlarm> reportActionAlarms = reportActionAlarmRepository
+                .findByMember(member, pageRequest);
+
+        return reportActionAlarms.stream()
+                .map(ReportActionAlarmResponse::from)
+                .toList();
+    }
+
+    public ReportActionResponse getReportActionAlarm(final Long reportActionAlarmId, final Member member) {
+        final ReportActionAlarm reportActionAlarm = reportActionAlarmRepository
+                .findByIdAndMember(reportActionAlarmId, member)
+                .orElseThrow(() -> new NotFoundException(ReportActionAlarmExceptionType.NOT_FOUND));
+
+        return ReportActionResponse.from(reportActionAlarm);
     }
 
 }
